Refactor lib building system
We only need one target: node without experimental properties. Since we’re not using any
backend/frontend related dependencies, this should be perfectly suited to making the
lib work with CRA and node.

(change) Refactor the way we build the library for npm. Should be cleaner now.
