+++ .babelrc
@@ -1,43 +1,13 @@
 {
-  "env": {
-    "development": {
-      "presets": [
-        ["@babel/preset-env", {
-          "targets": {
-            "node": "6.10"
-          }
-        }]
-      ],
-      "plugins": [
-        "transform-class-properties"
-      ]
-    },
-    "browser": {
-      "presets": [
-        ["@babel/preset-env", {
-          "targets": {
-            "browsers": "last 2 versions, ie 10-11"
-          },
-          "modules": false
-        }]
-      ],
-      "plugins": [
-        "transform-class-properties"
-      ]
-    },
-    "module": {
-      "presets": [
-        ["@babel/preset-env", {
-          "targets": {
-            "node": "6.10"
-          },
-          "modules": false
-        }]
-      ],
-      "plugins": [
-        "transform-class-properties"
-      ]
-    }
-  },
-  "sourceMaps": true
+  "presets": [
+    ["@babel/preset-env", {
+      "targets": {
+        "node": "10.12"
+      }
+    }]
+  ],
+  "plugins": [
+    "add-module-exports",
+    "transform-class-properties"
+  ]
 }
