--- .npmignore
@@ -1,4 +1,0 @@
-tests
-yarn.lock
-.npmignore
-.gitignore
