+++ package.json
@@ -2,17 +2,16 @@
   "name": "typed-models",
   "version": "1.5.0",
   "description": "",
-  "main": "./dist/index.js",
-  "browser": "./lib/index.js",
-  "module": "./module/index.js",
+  "main": "./lib/index.js",
+  "files": [
+    "/lib"
+  ],
   "scripts": {
+    "clean": "rm -rf ./lib",
     "test": "jest ./tests",
     "lint": "eslint",
     "codecov": "codecov",
-    "build:browser": "cross-env BABEL_ENV=browser babel ./src --out-dir ./lib --source-maps --copy-files",
-    "build:module": "cross-env BABEL_ENV=module babel ./src --out-dir ./module --source-maps --copy-files",
-    "build:node": "babel ./src --out-dir ./dist --source-maps --copy-files",
-    "build": "npm run build:node && npm run build:browser && npm run build:module",
+    "build": "yarn clean && babel src --out-dir lib",
     "release": "yarn publish --non-interactive"
   },
   "author": "Mateusz Klos",
@@ -20,10 +19,6 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/novopl/json-models.git"
-  },
-  "directories": {
-    "lib": "./",
-    "test": "./test"
   },
   "jest": {
     "testEnvironment": "node",
@@ -86,13 +81,13 @@
     "@babel/cli": "^7.11.6",
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
+    "babel-plugin-add-module-exports": "^1.0.4",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-preset-minify": "^0.5.1",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "codecov": "^3.7.2",
-    "cross-env": "^7.0.2",
-    "eslint": "^7.7.0",
-    "faker": "^5.0.0",
-    "jest": "^26.4.2"
+    "eslint": "^7.9.0",
+    "jest": "^26.4.2",
+    "codecov": "^3.7.2"
   }
 }
